bot/requirements.txt
python-telegram-bot>=20.0
python-dotenv
httpx
reportlab
